+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### ✨ Improved
 
 * Take AG darks during the pre-observing task.
+* Add `retry_without_parking` option to the shutdown recipe.
 
 ### 🔧 Fixed